feat: handle __resolveReference from federated schemas
In a federated schema you specify the `@keys` of an object. These keys are always available, regardless of where the query comes from. You can also (optionally) specify a `__resolveReference` function to take the keys and add in all the other missing properties. This function is required if not all properties are listed as part of the keys. This differs from all the other resolvers, so it merrits a special case.

Once https://github.com/dotansimha/graphql-code-generator/pull/2534 lands, I should be able to create an example that uses this.