fix: 🐛 Allow non-required fields on query return types directly
I keep running into this issue with our GraphQL servies - before hand
this schema would have tried to force me to put something in `index.ts`
to resolve the optional `notRequiredValue` field on `PartialReturnValue`
- as in the types it would mismatch, for some reason it comes through as
a `Promise`. I'm not sure if this is the best way to fix this, or if
there's some other work around/thing I'm meant to be doing that means
this wasn't necessary. But if others have run into this bug, and aren't
aware of any other ways to fix it - then this would be a good way to go
about it I think. It took far too much of my time arriving at this
solution - good luck to anyone wanting to figure out this codebase.
:smile: